add some size_t to get ride of gcc 4.9 warning

+++ src/base/range.h
@@ -103,7 +103,7 @@ class Range {
 template<class T>
 Range<T> Range<T>::EvenDivide(size_t n, size_t i) const {
   CHECK(valid());
-  CHECK_GT(n, 0);
+  CHECK_GT(n, (size_t)0);
   CHECK_LT(i, n);
   auto itv = static_cast<long double>(end_ - begin_) /
              static_cast<long double>(n);
